// src/app/core/models/maintenance/filter/intervention-maintenance-filter.model.ts
export class InterventionMaintenanceFilter {
  num_demande: string;
  date_demande: Date;
  statut: string;
  codeImmat: number;
  tonnage: number;
  activity: string;
  type_intervention: string;
  type_panne: number;
  demandeur: number;
  date_prise: Date;
  date_prevu: Date;
  date_fin: Date;
}